+++ packages/interop/deco-samples/test/cosmetics.test.js
@@ -1,7 +1,7 @@
 import { tableToSamples }       from '@analys/convert'
-import { TableCollection }      from '@foba/table'
-import { FRESH, METRO, SUBTLE } from '@palett/presets'
-import { says }                 from '@spare/says'
+import { TableCollection }            from '@foba/table'
+import { FRESH, METRO, PINE, SUBTLE } from '@palett/presets'
+import { says }                       from '@spare/says'
 import { test }                 from 'node:test'
 import { decoSamples }          from '../index.js'
 
@@ -9,6 +9,8 @@ const samples = tableToSamples(TableCollection.AeroEngineSpecs)
 test('cosmetics', () => {
   says['AeroEngineSpecs'](decoSamples(samples, {
     indexed: true,
-    indent: 1
-  }, SUBTLE, { str: FRESH, pos: METRO, neg: SUBTLE }))
+    indent: 1,
+    pres: { str: FRESH, pos: METRO, neg: SUBTLE },
+    key: PINE,
+  }))
 })