+++ packages/interop/deco-table/test/aoe.table.test.js
@@ -9,8 +9,8 @@ const AoEIIUnitsAttackByStages = {
     [ 9.8, 10, 10, 31.3 ],
     [ 6.7, 4.7, 7.3, 18 ],
     [ 4.5, 3, 5, NaN ],
-    [ 4, NaN, NaN, NaN ]
-  ]
+    [ 4, NaN, NaN, NaN ],
+  ],
 }
 
 
@@ -21,8 +21,8 @@ test('aoe table', () => {
     fill: ' ',
     ansi: true,
     pres: { pos: BESQUE, neg: OCEAN, str: METRO },
-    keys: { pos: BESQUE, neg: OCEAN, str: SUBTLE }
+    keys: { pos: BESQUE, neg: OCEAN, str: SUBTLE },
   }
 
-  console.log(decoTable(AoEIIUnitsAttackByStages, CONF, 2))
+  console.log(decoTable(AoEIIUnitsAttackByStages, 0, 2))
 })