+++ packages/interop/csv/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @spare/csv
 
+## 0.16.10
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/table-margin@0.16.10
+  - @spare/table-padder@0.16.10
+
+## 0.16.9
+
+### Patch Changes
+
+- update interface
+- Updated dependencies
+  - @spare/table-margin@0.16.9
+  - @spare/table-padder@0.16.9
+
 ## 0.16.8
 
 ### Patch Changes