+++ packages/interop/deco-samples/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @spare/deco-samples
 
+## 0.16.10
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @spare/deco-flat@0.16.10
+  - @spare/enum-brackets@0.16.10
+  - @spare/preset-deco@0.16.10
+  - @spare/node@0.16.10
+  - @spare/deco-config@0.16.10
+
+## 0.16.9
+
+### Patch Changes
+
+- update interface
+- Updated dependencies
+  - @spare/enum-brackets@0.16.9
+  - @spare/preset-deco@0.16.9
+  - @spare/deco-flat@0.16.9
+  - @spare/node@0.16.9
+  - @spare/deco-config@0.16.9
+
 ## 0.16.8
 
 ### Patch Changes